updated AddTask.tsx file for the handlekeydown event, TaskStore incl addTask, getTotalTasks and getCompletedTasks
